Implement .env support and hot-reloading
Load environment variables from .env and .env.local, and also watch these files
for changes.

Java does not normally allow you to change the environment variables of the
running JVM, and we go through some serious shenanigans to support it.

We also unified the watching of deps.edn and .env so we don't have to start so
many different beholder watchers. Other launchpad extensions can plug in to this
mechanism as well to watch additional files.