Correct dotenv watch semantics and refactor
When using the dotenv functionality I have noticed that we were not
correctly handling file creation and deletion and decided to fix it.

This patch makes sure that we do the following no matter the initial
state (aka, it does not matter if the files are missing when you launch
the pad):

  | .env  | .env.local | Values From  |
  | ----- |----------- | ------------ |
  | ✅    | ✅         | `.env.local` |
  | ✅    | ❌         | `.env`       |
  | ❌    | ✅         | `.env.local` |
  | ❌    | ❌         | latest state |

The last line means we currently do not keep track and therefore cannot
remove already set environment variables.

The patch also refactors the code a bit to remove code duplication:

  * `lamdbaisland.launchpad.env` is babashka-compatible and can is
     required into `lambdaistland.launchpad` in order to populate
     `:env`.

  * `lamdbaisland.launchpad.env.hacks`: is now the one evaluated by the
    watcher, uses `com.lamdbaisland.classpath` and contains the
    "nitty-gritty" details.